- Sweet component installed
- Alerts Added to show messages to the user.
- Added  a properly handle when multiple reminders appear on the same date.

+++ src/app/pages/reminder/reminder.component.ts
@@ -6,6 +6,7 @@ import { ReminderService } from 'src/app/services/reminder.service';
 import { Reminder } from 'src/app/models/reminder.model';
 
 import * as moment from 'moment';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-reminder',
@@ -33,29 +34,49 @@ export class ReminderComponent implements OnInit {
   }
 
   save(f: NgForm) {
-    if (!f.pristine)
-      alert('Test pristine');
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      showConfirmButton: false,
+      timer: 3000
+    })
+
     this.reminder.DateTime = new Date(f.controls['date'].value + ' ' + f.controls['time'].value);
     if (this.reminder.Id == 0) {
-      this.reminderService.addReminder(this.reminder); 
+      this.reminderService.addReminder(this.reminder);
+      Toast.fire({icon: 'success',title: 'Reminder created successfully'});
     }
-    else{      
+    else {
       this.reminderService.editReminder(this.reminder);
+      Toast.fire({icon: 'success',title: 'Reminder saved successfully'});
     }
     this.onCloseEdit.emit();
     this.onSave.emit();
   }
 
-  delete(f){
-    this.reminderService.deleteReminder(this.reminder);
-    this.onClose.emit();
+  delete() {
+    Swal.fire({
+      title: 'Are you sure to delete this reminder?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.reminderService.deleteReminder(this.reminder);
+        this.onSave.emit();
+        this.onClose.emit();
+      }
+    });
   }
-  
+
   cancel(f: NgForm) {
     if (this.reminder.Id == 0) {
       this.onClose.emit();
     }
-    else{      
+    else {
       this.onCloseEdit.emit(false);
     }
   }
